clean up tdvp module structure
