+++ docs/examples/simulation_parameters.md
@@ -183,7 +183,9 @@ Digital circuit simulation on an MPS defaults to **`gate_mode="mpo"`** (generic 
 
 Long-range entangling gates in TDVP mode automatically apply **gate-local bond support retention** on anchor bonds crossed by the active two-qubit gate (handled inside the digital TDVP path; not a user-facing parameter). The rule enforces an effective minimum bond dimension of `min(2, max_bond_dim)` when a finite cap is set, or `2` when `max_bond_dim` is unset. Retention never exceeds `max_bond_dim`.
 
-Use **`tdvp_sweeps`** (default `1`) to split each TDVP evolution step into multiple substeps of equal total time. Higher values can improve accuracy at higher cost. The setting applies to all TDVP kernels (two-site, single-site, local dynamic, and global dynamic TDVP) on `AnalogSimParams`, `StrongSimParams`, and `WeakSimParams`.
+Use **`tdvp_sweeps`** (default `1`) to split each TDVP evolution step into multiple substeps of equal total time. Higher values can improve accuracy at higher cost. The setting applies to all TDVP kernels on `AnalogSimParams`, `StrongSimParams`, and `WeakSimParams`.
+
+Use **`tdvp_mode`** (default `"dynamic"`) to select the TDVP integrator: `"1site"` (1TDVP), `"2site"` (2TDVP), or `"dynamic"` (adaptive single/two-site updates). Long-range digital gate support requires `"dynamic"`.
 
 Substep geometry: each substep is **symmetric** (left-to-right then right-to-left) at evolution time `step_time / tdvp_sweeps` for analog (`dt`) and digital gates. The total generator time applied to one digital gate remains `1` across all substeps. Noise and dissipation after TDVP still use the full physical step `dt` in analog simulation.
