+++ src/mqt/yaqs/core/methods/tdvp/__init__.py
@@ -1,0 +1,22 @@
+# Copyright (c) 2025 - 2026 Chair for Design Automation, TUM
+# All rights reserved.
+#
+# SPDX-License-Identifier: MIT
+#
+# Licensed under the MIT License
+
+"""Time-dependent variational principle (TDVP) methods.
+
+Public entry point: :func:`mqt.yaqs.core.methods.tdvp.tdvp.tdvp`.
+Low-level kernels live in :mod:`mqt.yaqs.core.methods.tdvp.primitives`.
+Sweep integrators live in :mod:`mqt.yaqs.core.methods.tdvp.integrators`.
+Sweep helpers live in :mod:`mqt.yaqs.core.methods.tdvp.sweep_utils`.
+Long-range digital gate bond support lives in
+:mod:`mqt.yaqs.core.methods.tdvp.bond_support`.
+"""
+
+from __future__ import annotations
+
+from .tdvp import tdvp
+
+__all__ = ["tdvp"]
